MB-63171: Temporarily lock query to earlier SHA
Query merged from cypher->master prior to corresponding changes in
tlm. Those tlm changes can't be merged yet due to coordinating changes
in backup and couchbase-cli that need to be vetted by that team. To
fix build breakage, locking query to the last `master` SHA prior to
the merge of the standalone-tools changes.

Change-Id: I7abc59483f715fd914644d6910de4de11d81bb39
Reviewed-on: https://review.couchbase.org/c/manifest/+/216145
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>