MB-64191: Update the revision number of libcouchbase to refs/tags/3.3.14
We must use the libcouchbase fix pushed in CCBC-1653 to address the
libcouchbase bug reported in the CBSE-18371, wherein earlier
libcouchbase was not honouring the timeout provided by Eventing for the
analytics queries and was defaulting to a 75-second timeout.

The libcouchbase fix for this issue is pushed in CCBC-1653 and is part
of the libcouchbase release version `3.3.14`. Hence, we are updating the
revision of libcouchbase to `refs/tags/3.3.14` in the `trinity.xml`
manifest file targeting Couchbase Server 7.6.5.

Change-Id: I3a0426d852f184d0a4018fef167345b3ad2409f8
Reviewed-on: https://review.couchbase.org/c/manifest/+/219954
Tested-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>