MB-64191: Update the revision number of libcouchbase to refs/tags/3.3.14
We must use the libcouchbase fix pushed in CCBC-1653 to address the
libcouchbase bug reported in the CBSE-18371, wherein earlier
libcouchbase was not honouring the timeout provided by Eventing for the
analytics queries and was defaulting to a 75-second timeout.

The libcouchbase fix for this issue is pushed in CCBC-1653 and is part
of the libcouchbase release version `3.3.14`. Hence, we are updating the
revision of libcouchbase to `refs/tags/3.3.14` in the `morpheus.xml`
manifest file targeting Couchbase Server 8.0.0.

Change-Id: I520de81a0be17fdf0f5e27b08174ac3def86747f
Reviewed-on: https://review.couchbase.org/c/manifest/+/219960
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>