MB-64191: Update the revision number of libcouchbase to refs/tags/3.3.14
We must use the libcouchbase fix pushed in CCBC-1653 to address the
libcouchbase bug reported in the CBSE-18371, wherein earlier
libcouchbase was not honouring the timeout provided by Eventing for the
analytics queries and was defaulting to a 75-second timeout.

The libcouchbase fix for this issue is pushed in CCBC-1653 and is part
of the libcouchbase release version `3.3.14`. Hence, we are updating the
revision of libcouchbase to `refs/tags/3.3.14` in the
`branch-master.xml` manifest file targeting Couchbase Server 0.0.0.

Change-Id: I4c49554ffff67412c3c9a7722af2b71d559afa42
Reviewed-on: https://review.couchbase.org/c/manifest/+/219961
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>